Add Multus with Flannel and SRIOV CNI support
This commit provides Kubernetes networking support for Multus
with SRIOV CNI support both on arm64 and amd64. A special
configuration file for Broadcom smartNIC Stingray PS225 is
provided as an example.

So it can not only be used as a sample container networking
support for Stingray PS225 by its interfaces configured with
VFs, but also be used as a generic SRIOV support for
various ethernet NICs.

Here the Flannel would be provided as the default CNIs for
any pods without explicit annotations.

For detailed information, please refer the README.md in the
commit.

Signed-off-by: Trevor Tao <trevor.tao@arm.com>
Change-Id: I6ddfaa8f7554a40bfe4974206da8cc45f240b6bc
Signed-off-by: Trevor Tao <trevor.tao@arm.com>