Add Multus with Calico and SRIOV CNI support
This commit provides Kubernetes networking support for Multus
with SRIOV CNI/Calico CNI support both on arm64 and amd64.

A special configuration file for Broadcom smartNIC Stingray
PS225 is provided as an example.

Here the Calico would be provided as the default CNIs for
any pods without explicit annotations.

Updated the README.md to add more explanations for current
work and reflect the change.

For detailed information, please refer the README.md in the
commit.

Signed-off-by: Trevor Tao <trevor.tao@arm.com>
Change-Id: I60ac7d636be8e272cd82c3697833947850e7f49e
Signed-off-by: Trevor Tao <trevor.tao@arm.com>