Optimize 2 points on the installation for CNI
Optimize the installation of k8s CNI at 2 sides:

1. Change the parameter order for the most needed
parameters first as the following:
    CNI_TYPE
    POD_NETWORK_CIDR
    K8s_MASTER_IP
    SERVICE_CIDR
    CLUSTER_IP
    DEV_NAME

so that when installing manually, less parameters need to
be input because the default parameters are enough.

2. Taint the master node first for the case there is just
the master node to use. So the daemon pods of a CNI could
be scheduled on the master. If not, the installation for
the CNI may fail due to no node to schedule the pods.

The user could drain or set NoSchedule to close the master
for scheduling if needed:
kubectl drain <master node name>
or:
kubectl taint nodes <master node name> key=value:NoSchedule

Signed-off-by: Trevor Tao <trevor.tao@arm.com>
Change-Id: I8198814af22e687ccb223b9c2b767e64943588a7
Signed-off-by: Trevor Tao <trevor.tao@arm.com>