Update ponsim install to cord-7.0-arm64
With this new version, install.sh will basically install SiaB with
PONSim. To do this, it will use our own version of automation-tools,
which internally will use the helm-charts on version cord-7.0-arm64.

The helm-charts are expected to be modified such that the deployment
will support IEC Type 2 3-node K8S clusters.

Signed-off-by: Ciprian Barbu <ciprian.barbu@enea.com>
Change-Id: I459f136b0d97c0ad9fecddd8e03187ee8c7d5237