Mount results dir for SEBA robot tests
SEBA validation is run from a Docker container but we need the robot
reports and logs to be avilable on the host (jenkins slave) to be
uploaded after the job finishes. This patch mounts a results dir for
this purpose.
Also the testing scripts have been refactored and cleaned up.

Fixes IEC-45

Signed-off-by: Ciprian Barbu <ciprian.barbu@enea.com>
Change-Id: I26b7162fac58b347a5ea5f529e9368c1bf09b083